refactor: separated the searchbars and togglebuttons into its own classes
