+++ lib/pages/farm/desktop_farm.dart
@@ -244,7 +244,6 @@ class _DesktopFarmState extends State<DesktopFarm> {
         myController: myController,
         layoutHgt: layoutHgt,
         layoutWdt: layoutWdt,
-        bloc: bloc,
       );
     } else {
       return WebSearchBar(
@@ -252,205 +251,7 @@ class _DesktopFarmState extends State<DesktopFarm> {
         myController: myController,
         layoutHgt: layoutHgt,
         layoutWdt: layoutWdt,
-        bloc: bloc,
       );
     }
-  }
-}
-
-class ToggleFarmButton extends StatelessWidget {
-  const ToggleFarmButton({
-    super.key,
-    required this.layoutWidth,
-    required this.layoutHeight,
-    required TextEditingController myController,
-  }) : _farmTextController = myController;
-
-  final double layoutWidth;
-  final double layoutHeight;
-  final TextEditingController _farmTextController;
-
-  @override
-  Widget build(BuildContext context) {
-    final isAllFarmsSelection =
-        context.select((FarmBloc bloc) => bloc.state.isAllFarms);
-    final isWeb =
-        kIsWeb && (MediaQuery.of(context).orientation == Orientation.landscape);
-    return Container(
-      width: isWeb ? 200 : layoutWidth,
-      height: 40,
-      decoration: boxDecoration(Colors.grey[900]!, 100, 1, Colors.grey[400]!),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
-        children: [
-          Container(
-            width: isWeb ? 90 : (layoutWidth / 2) - 5,
-            decoration: isAllFarmsSelection
-                ? boxDecoration(
-                    Colors.grey[600]!,
-                    100,
-                    0,
-                    Colors.transparent,
-                  )
-                : boxDecoration(
-                    Colors.transparent,
-                    100,
-                    0,
-                    Colors.transparent,
-                  ),
-            child: TextButton(
-              onPressed: () {
-                if (!isAllFarmsSelection) {
-                  _farmTextController.clear();
-                  context
-                      .read<FarmBloc>()
-                      .add(OnChangeFarmTab(isAllFarms: true));
-                }
-              },
-              child: Text(
-                'All Farms',
-                style: textStyle(
-                  Colors.white,
-                  16,
-                  isBold: true,
-                  isUline: false,
-                ),
-              ),
-            ),
-          ),
-          Container(
-            width: isWeb ? 90 : (layoutWidth / 2) - 5,
-            decoration: isAllFarmsSelection
-                ? boxDecoration(
-                    Colors.transparent,
-                    100,
-                    0,
-                    Colors.transparent,
-                  )
-                : boxDecoration(
-                    Colors.grey[600]!,
-                    100,
-                    0,
-                    Colors.transparent,
-                  ),
-            child: TextButton(
-              onPressed: () {
-                final isWalletConnected =
-                    context.read<WalletBloc>().state.isWalletConnected;
-                if (isWalletConnected) {
-                  if (isAllFarmsSelection) {
-                    _farmTextController.clear();
-                    context
-                        .read<FarmBloc>()
-                        .add(OnChangeFarmTab(isAllFarms: false));
-                  }
-                } else {
-                  context.showWalletWarningToast();
-                }
-              },
-              child: Text(
-                'My Farms',
-                style:
-                    textStyle(Colors.white, 16, isBold: true, isUline: false),
-              ),
-            ),
-          )
-        ],
-      ),
-    );
-  }
-}
-
-class WebSearchBar extends StatelessWidget {
-  const WebSearchBar({
-    super.key,
-    required this.isWeb,
-    required this.myController,
-    required this.layoutWdt,
-    required this.layoutHgt,
-    required this.bloc,
-  });
-
-  final bool isWeb;
-  final TextEditingController myController;
-  final double layoutWdt;
-  final double layoutHgt;
-  final FarmBloc bloc;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: isWeb ? 250 : layoutWdt / 2,
-      height: isWeb ? 40 : layoutHgt * 0.05,
-      decoration: boxDecoration(Colors.grey[900]!, 100, 1, Colors.grey[300]!),
-      child: Row(
-        children: [
-          const SizedBox(width: 8),
-          const Icon(Icons.search, color: Colors.white),
-          const SizedBox(width: 8),
-          Expanded(
-            child: TextFormField(
-              controller: myController,
-              onChanged: (value) {
-                bloc.add(OnSearchFarms(searchedName: value));
-              },
-              decoration: const InputDecoration(
-                border: InputBorder.none,
-                contentPadding: EdgeInsets.only(bottom: 8.5),
-                hintText: 'Search a farm',
-                hintStyle: TextStyle(color: Colors.white),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class MobileSearchBar extends StatelessWidget {
-  const MobileSearchBar({
-    super.key,
-    required this.isWeb,
-    required this.myController,
-    required this.layoutWdt,
-    required this.layoutHgt,
-    required this.bloc,
-  });
-
-  final bool isWeb;
-  final TextEditingController myController;
-  final double layoutWdt;
-  final double layoutHgt;
-  final FarmBloc bloc;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: isWeb ? 250 : layoutWdt / 2,
-      height: isWeb ? 40 : layoutHgt * 0.05,
-      decoration: boxDecoration(Colors.grey[900]!, 100, 1, Colors.grey[300]!),
-      child: Row(
-        children: [
-          const SizedBox(width: 8),
-          const Icon(Icons.search, color: Colors.white),
-          const SizedBox(width: 8),
-          Expanded(
-            child: TextFormField(
-              controller: myController,
-              onChanged: (value) {
-                bloc.add(OnSearchFarms(searchedName: value));
-              },
-              decoration: InputDecoration(
-                border: InputBorder.none,
-                contentPadding: EdgeInsets.only(bottom: layoutHgt * 0.022),
-                hintText: 'Search a farm',
-                hintStyle: const TextStyle(color: Colors.white),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
   }
 }
