+++ lib/pages/farm/widgets/widgets.dart
@@ -4,8 +4,10 @@ export 'loading_error.dart';
 export 'my_farm_item.dart';
 export 'no_data.dart';
 export 'no_wallet.dart';
+export 'search_bar.dart';
 export 'single_logo_farm_title.dart';
 export 'sport_token.dart';
 export 'stake_approve_button.dart';
+export 'toggle_farm_button.dart';
 export 'unstake_approve_button.dart';
 export 'unsupported_chain.dart';
